Revert "Add action for setting selection of nodes and edges"
This reverts commit c05963f055f087e7bc8a15e3074a1d020af7c542
in favor of re-getting the solution view every time we access the
solution property. It is simpler to implement to start, since otherwise
we need to intercept all places where we could set the solution
attribute (add node, add edge, set feature values). We can solve the
efficiency problem separately if/when needed.