refactor: :art: Remove all references to tracks.time_attr and tracks.pos_attr
This was done with the minimal disruption to the existing code. Therefore, I used feature.key to get the same strings that were used as keys in the actions before. A larger refactor will be needed to actually use the features as keys instead of strings, if we want to do that in the future.