feat: :construction: Propose implementation of the RegionpropsAnnotator
I changed the signature of the base annotator to take the Tracks at initialization (since we need to know if it's 2d or 3d to determine valid features). I also changed the signature of compute and update to not take the Tracks, and to compute for all features. We still need to find a way to let the user pick which regionprops features to compute, test them more thoroughly, and see if the API makes sense for other Annotators.