feat: :sparkles: Add option to force remove merge edges when adding new edges
This will allow us to promt the user when there is an invalid action, and then re-submit the action with the "force" option. It also sets a framework for doing this type of thing more generally, and having an "always force" setting in finn if the user wants to stop being prompted.